refactor: Get ride of the StateT monad
The StateT monad was not really useful, and it was even harmful, as the
context obtained in the `recover*` functions has the complete stack, up
until the main. This does not make sense, because the calling function
should not care about the context it has been called.

This commit fixes #1.